+++ lib/util/extensions/game_component_extensions.dart
@@ -269,6 +269,7 @@ extension GameComponentExtensions on GameComponent {
     double end = 1.0,
     Curve curve = Curves.linear,
     bool autoStart = true,
+    bool infinite = false,
     VoidCallback? onFinish,
     ValueChanged<double>? onChange,
   }) {
@@ -280,8 +281,9 @@ extension GameComponentExtensions on GameComponent {
       onFinish: onFinish,
       onChange: onChange,
       autoStart: autoStart,
+      infinite: infinite,
     );
-    gameRef.add(valueGenerator);
+    add(valueGenerator);
     return valueGenerator;
   }