+++ lib/util/value_generator_component.dart
@@ -13,10 +13,12 @@ class ValueGeneratorComponent extends Component {
   final VoidCallback? onFinish;
   final ValueChanged<double>? onChange;
   final bool autoStart;
+  final bool infinite;
 
   double _currentValue = 0;
   double _displacement = 0;
   bool _isRunning = false;
+  bool _reversing = false;
 
   ValueGeneratorComponent(
     this.duration, {
@@ -25,7 +27,8 @@ class ValueGeneratorComponent extends Component {
     this.curve = Curves.decelerate,
     this.onFinish,
     this.onChange,
-    this.autoStart = false,
+    this.autoStart = true,
+    this.infinite = false,
   }) {
     _isRunning = autoStart;
     _displacement = end - begin;
@@ -35,9 +38,21 @@ class ValueGeneratorComponent extends Component {
   void updateTree(double dt) {
     if (!_isRunning) return;
 
-    _currentValue += dt * _maxInMilliSeconds;
+    if (_reversing) {
+      _currentValue -= dt * _maxInMilliSeconds;
+    } else {
+      _currentValue += dt * _maxInMilliSeconds;
+    }
+
     if (_currentValue >= duration.inMilliseconds) {
-      finish();
+      if (infinite) {
+        _reversing = true;
+      } else {
+        finish();
+      }
+    } else if (_currentValue <= 0 && _reversing) {
+      _currentValue = 0;
+      _reversing = false;
     } else {
       double value = curve.transform(_currentValue / duration.inMilliseconds);
       double realValue = begin + (_displacement * value);